+++ frontend/src/components/RecipeNoImgDetailedDialog.tsx
@@ -7,10 +7,11 @@ import ConfirmationDialog from "./ConfirmationDialog"
 interface RecipeDetailDialogProps {
     recipe: RecipeModel,
     onDismiss: () => void,
-    onEdit: () => void,
-    onDelete: () => void
+    onEdit?: () => void,
+    onDelete?: () => void
     createdAtString: string,
     updatedAtString: string,
+    isPublic?: boolean
 }
 
 const RecipeNoImgDetailDialog: React.FC<RecipeDetailDialogProps> = ({
@@ -19,7 +20,8 @@ const RecipeNoImgDetailDialog: React.FC<RecipeDetailDialogProps> = ({
                                                                         onEdit,
                                                                         onDelete,
                                                                         createdAtString,
-                                                                        updatedAtString
+                                                                        updatedAtString,
+                                                                        isPublic
                                                                     }) => {
     const [showConfirmDelete, setShowConfirmDelete] = useState(false);
 
@@ -39,20 +41,28 @@ const RecipeNoImgDetailDialog: React.FC<RecipeDetailDialogProps> = ({
                         {updatedAtString}
                     </p>
                     <Button variant="secondary" onClick={onDismiss}>Close</Button>
-                    <Button variant="primary" onClick={onEdit}>Edit</Button>
-                    <Button variant="danger" onClick={() => setShowConfirmDelete(true)}>Delete</Button>
+                    {!isPublic &&
+                        <>
+                            <Button variant="primary" onClick={onEdit}>Edit</Button>
+                            <Button variant="danger" onClick={() => setShowConfirmDelete(true)}>Delete</Button>
+                        </>
+                    }
+
                 </Modal.Footer>
             </Modal>
-            <ConfirmationDialog
-                show={showConfirmDelete}
-                title="Confirm Delete"
-                message="Are you sure you want to delete this recipe?"
-                onConfirm={() => {
-                    onDelete();
-                    setShowConfirmDelete(false);
-                }}
-                onCancel={() => setShowConfirmDelete(false)}
-            />
+            {!isPublic &&
+                <ConfirmationDialog
+                    show={showConfirmDelete}
+                    title="Confirm Delete"
+                    message="Are you sure you want to delete this recipe?"
+                    onConfirm={() => {
+                        if (onDelete) {
+                            onDelete();
+                        }
+                        setShowConfirmDelete(false);
+                    }}
+                    onCancel={() => setShowConfirmDelete(false)}/>
+            }
         </div>
     );
 }
